requirements.txt
exceptiongroup==1.1.3
importlib-metadata==6.7.0
iniconfig==2.0.0
mysql-connector-python==8.0.33
packaging==23.2
pluggy==1.2.0
protobuf==3.20.3
pytest==7.4.3
python-dotenv==0.21.1
tomli==2.0.1
typing-extensions==4.7.1
zipp==3.15.0
